fixed perspective mapview, now with movement
git-svn-id: https://qtlobby.googlecode.com/svn/trunk@465 5a83989a-60d1-11de-a6a3-193033c112f2

+++ src/MapRendererWidget.cpp
@@ -72,7 +72,7 @@ void MapRendererWidget::resizeGL(int w, int h) {
     glLoadIdentity();
 
     const float angle=30;
-    gluPerspective(angle,w/h,1,1000);
+    gluPerspective(angle,double(w)/(double(h),1,1000);
 
     //if (w <= h)
     //    glOrtho (dx+lastZoom*-100, dx+lastZoom*100,/*left,right*/
